+++ dart/lib/src/event_processor/enricher/io_platform_memory.dart
@@ -6,32 +6,47 @@ import '../../sentry_options.dart';
 // Get total & free platform memory (in bytes) for linux and windows operating systems.
 // Source: https://github.com/onepub-dev/system_info/blob/8a9bf6b8eb7c86a09b3c3df4bf6d7fa5a6b50732/lib/src/platform/memory.dart
 class PlatformMemory {
-  PlatformMemory(this.options);
+  PlatformMemory(this.options) {
+    if (options.platformChecker.platform.isWindows) {
+      // Check for WMIC (deprecated in newer Windows versions)
+      // https://techcommunity.microsoft.com/blog/windows-itpro-blog/wmi-command-line-wmic-utility-deprecation-next-steps/4039242
+      useWindowsWmci =
+          File('C:\\Windows\\System32\\wbem\\wmic.exe').existsSync();
+      if (!useWindowsWmci) {
+        useWindowsPowerShell = File(
+                'C:\\Windows\\System32\\WindowsPowerShell\\v1.0\\powershell.exe')
+            .existsSync();
+      } else {
+        useWindowsPowerShell = false;
+      }
+    } else {
+      useWindowsWmci = false;
+      useWindowsPowerShell = false;
+    }
+  }
 
   final SentryOptions options;
+  late final bool useWindowsWmci;
+  late final bool useWindowsPowerShell;
 
-  int? getTotalPhysicalMemory() {
+  Future<int?> getTotalPhysicalMemory() async {
     if (options.platformChecker.platform.isLinux) {
       return _getLinuxMemInfoValue('MemTotal');
     } else if (options.platformChecker.platform.isWindows) {
-      return _getWindowsWmicValue('ComputerSystem', 'TotalPhysicalMemory');
+      if (useWindowsWmci) {
+        return _getWindowsWmicValue('ComputerSystem', 'TotalPhysicalMemory');
+      } else if (useWindowsPowerShell) {
+        return _getWindowsPowershellTotalMemoryValue();
+      } else {
+        return null;
+      }
     } else {
       return null;
     }
   }
 
-  int? getFreePhysicalMemory() {
-    if (options.platformChecker.platform.isLinux) {
-      return _getLinuxMemInfoValue('MemFree');
-    } else if (options.platformChecker.platform.isWindows) {
-      return _getWindowsWmicValue('OS', 'FreePhysicalMemory');
-    } else {
-      return null;
-    }
-  }
-
-  int? _getWindowsWmicValue(String section, String key) {
-    final os = _wmicGetValueAsMap(section, [key]);
+  Future<int?> _getWindowsWmicValue(String section, String key) async {
+    final os = await _wmicGetValueAsMap(section, [key]);
     final totalPhysicalMemoryValue = os?[key];
     if (totalPhysicalMemoryValue == null) {
       return null;
@@ -43,12 +58,10 @@ class PlatformMemory {
     return size;
   }
 
-  int? _getLinuxMemInfoValue(String key) {
-    final meminfoList = _exec('cat', ['/proc/meminfo'])
-            ?.trim()
-            .replaceAll('\r\n', '\n')
-            .split('\n') ??
-        [];
+  Future<int?> _getLinuxMemInfoValue(String key) async {
+    final result = await _exec('cat', ['/proc/meminfo']);
+    final meminfoList =
+        result?.trim().replaceAll('\r\n', '\n').split('\n') ?? [];
 
     final meminfoMap = _listToMap(meminfoList, ':');
     final memsizeResults = meminfoMap[key]?.split(' ') ?? [];
@@ -65,11 +78,11 @@ class PlatformMemory {
     return memsize;
   }
 
-  String? _exec(String executable, List<String> arguments,
-      {bool runInShell = false}) {
+  Future<String?> _exec(String executable, List<String> arguments,
+      {bool runInShell = false}) async {
     try {
       final result =
-          Process.runSync(executable, arguments, runInShell: runInShell);
+          await Process.run(executable, arguments, runInShell: runInShell);
       if (result.exitCode == 0) {
         return result.stdout.toString();
       }
@@ -82,16 +95,16 @@ class PlatformMemory {
     return null;
   }
 
-  Map<String, String>? _wmicGetValueAsMap(String section, List<String> fields) {
+  Future<Map<String, String>?> _wmicGetValueAsMap(
+      String section, List<String> fields) async {
     final arguments = <String>[section];
     arguments
       ..add('get')
       ..addAll(fields.join(', ').split(' '))
       ..add('/VALUE');
 
-    final list =
-        _exec('wmic', arguments)?.trim().replaceAll('\r\n', '\n').split('\n') ??
-            [];
+    final result = await _exec('wmic', arguments);
+    final list = result?.trim().replaceAll('\r\n', '\n').split('\n') ?? [];
 
     return _listToMap(list, '=');
   }
@@ -107,5 +120,24 @@ class PlatformMemory {
       }
     }
     return map;
+  }
+
+  Future<int?> _getWindowsPowershellTotalMemoryValue() async {
+    final command =
+        'Get-CimInstance Win32_ComputerSystem | Select-Object -ExpandProperty TotalPhysicalMemory';
+
+    final result = await _exec('powershell.exe',
+        ['-NoProfile', '-NonInteractive', '-Command', command]);
+    if (result == null) {
+      return null;
+    }
+
+    final value = result.trim();
+    final size = int.tryParse(value);
+    if (size == null) {
+      return null;
+    }
+
+    return size;
   }
 }