+++ dart/test/contexts_test.dart
@@ -134,6 +134,19 @@ void main() {
       expect(clone['theme'], {'value': 'material'});
       expect(clone['version'], {'value': 9});
     });
+
+    test('set runtimes', () {
+      final contexts = Contexts();
+      contexts.runtimes = [
+        const SentryRuntime(name: 'testRT1', version: '1.0'),
+        const SentryRuntime(name: 'testRT2', version: '2.0'),
+      ];
+      expect(contexts.runtimes.length, 2);
+      expect(contexts.runtimes.first.name, 'testRT1');
+      expect(contexts.runtimes.first.version, '1.0');
+      expect(contexts.runtimes.last.name, 'testRT2');
+      expect(contexts.runtimes.last.version, '2.0');
+    });
   });
 
   group('parse contexts', () {